requirements.txt
pyyaml>=6.0
pre-commit
